+++ movingpandas/__init__.py
@@ -7,4 +7,4 @@ from .trajectory_collection import TrajectoryCollection
 from .trajectory_plotter import TrajectoryPlotter, TrajectoryCollectionPlotter
 
 name = 'movingpandas'
-__version__ = '0.1.dev2'
+__version__ = '0.2.rc2'
