### 2020-01-12 97b98d4
Add scikit-learn for notebooks
