+++ environment.yml
@@ -15,3 +15,4 @@ dependencies:
   - proj
   - cartopy
   - geoviews
+  - scikit-learn
