### 2020-01-12 8be2095
Add geopy for notebooks
