### 2020-01-15 #18 @anitagraser
Thank you @lwasser! I think GeoPandas developers would be a good fit.
