### 2020-01-16 #18 @martinfleis
@lwasser that seems to be doable. Count me in.
