### 2020-01-17 #18 @jlpalomino
## Editor checks:

- [X] **License:** The package has an OSI accepted license

---

## Editor comments

---

Reviewers: Ivan Ogasawara (xmnlab) and Martin Fleischmann (martinfleis)
Due date: February 7th, 2020
