### 2020-01-19 78c9dca
Fix shapely to 1.7a3
