### 2020-01-19 4be96cd
Pin shapely to 1.7a3 (#46)
* Fix shapely to 1.7a3

* ==
