### 2020-01-19 3b6fe65
Use Grouper in split_by_date
