### 2020-01-25 18703b5
Add TrajectoryCollectionAggregator
