+++ movingpandas/__init__.py
@@ -5,6 +5,7 @@
 from .trajectory import Trajectory
 from .trajectory_collection import TrajectoryCollection
 from .trajectory_plotter import TrajectoryPlotter, TrajectoryCollectionPlotter
+from .trajectory_aggregator import TrajectoryCollectionAggregator
 
 name = 'movingpandas'
 __version__ = '0.2.rc2'
