### 2020-01-25 f91b986
Add TrajectoryCollectionAggregator (#48)
