+++ movingpandas/trajectory_aggregator.py
@@ -1,0 +1,281 @@
+# -*- coding: utf-8 -*-
+
+import math
+from pandas import DataFrame
+from geopandas import GeoDataFrame
+from shapely.geometry import Point, LineString
+from shapely.ops import nearest_points
+
+from .geometry_utils import azimuth, angular_difference, measure_distance_spherical, measure_distance_euclidean, R_EARTH
+
+
+class TrajectoryCollectionAggregator:
+    def __init__(self, traj_collection, max_distance, min_distance, min_stop_duration, min_angle=45):
+        self.traj_collection = traj_collection
+        self.max_distance = max_distance
+        self.min_distance = min_distance
+        self.min_stop_duration = min_stop_duration
+        self.min_angle = min_angle
+        self.is_latlon = self.traj_collection.trajectories[0].is_latlon()
+        print('Extracting significant points ...')
+        self.significant_points = self.extract_significant_points()
+        print('  No. significant points: {}'.format(len(self.significant_points)))
+        print('Clustering significant points ...')
+        self.clusters = self.cluster_significant_points()
+        print('  No. clusters: {}'.format(len(self.clusters)))
+        print('Computing flows ...')
+        self.flows = self.compute_flows_between_clusters()
+        print('Flows ready!')
+
+    def get_significant_points_gdf(self):
+        if not self.significant_points:
+            self.extract_significant_points()
+        df = DataFrame(self.significant_points, columns=['geometry'])
+        return GeoDataFrame(df)
+
+    def extract_significant_points(self):
+        significant_points = []
+        for traj in self.traj_collection.trajectories:
+            a = PtsExtractor(traj, self.max_distance, self.min_distance, self.min_stop_duration, self.min_angle)
+            significant_points = significant_points + a.find_significant_points()
+        return significant_points
+
+    def cluster_significant_points(self):
+        points = self.significant_points
+        bbox = self.get_significant_points_gdf().total_bounds
+        cell_size = self.max_distance / 10
+        if self.is_latlon:
+            cell_size = cell_size / R_EARTH * 360
+        grid = Grid(bbox, cell_size)
+        grid.insert_points(points)
+        grid.redistribute_points(points)
+        return grid.resulting_clusters
+
+    def get_clusters_gdf(self):
+        if not self.clusters:
+            self.cluster_significant_points()
+        df = DataFrame([cluster.centroid for cluster in self.clusters], columns=['geometry'])
+        df['n'] = [len(cluster.points) for cluster in self.clusters]
+        return GeoDataFrame(df)
+
+    def compute_flows_between_clusters(self):
+        sg = SequenceGenerator(self.get_clusters_gdf(), self.traj_collection)
+        return sg.create_flow_lines()
+
+    def get_flows_gdf(self):
+        if not self.flows:
+            self.compute_flows_between_clusters()
+        return GeoDataFrame(self.flows)
+
+
+class PtsExtractor:
+    def __init__(self, traj, max_distance, min_distance, min_stop_duration, min_angle):
+        self.i = 0
+        self.j = 1
+        self.k = 2
+        self.traj = traj
+        self.n = self.traj.df.geometry.count()
+        self.max_distance = max_distance
+        self.min_distance = min_distance
+        self.min_stop_duration = min_stop_duration
+        self.min_angle = min_angle
+        self.significant_points = [self.traj.get_start_location(), self.traj.get_end_location()]
+
+    def find_significant_points(self):
+        while self.j < self.n - 1:
+            if self.is_representative_max_distance():
+                self.significant_points.append(self.traj.df.iloc[self.j].geometry)
+                self.i = self.j
+                self.j = self.i + 1
[…]
