### 2020-01-25 anitagraser/movingpandas#49 @martinfleis
Installation instructions
Hi,
this is a part of https://github.com/pyOpenSci/software-review/issues/18.

**Naive installation**
As a naive user, if I make a new python environment and do `pip install movingpandas`, the installation fails on Mac.

```
Collecting movingpandas
  Downloading movingpandas-0.2rc1.tar.gz (20 kB)
Collecting numpy
  Downloading numpy-1.18.1-cp38-cp38-macosx_10_9_x86_64.whl (15.2 MB)
     |████████████████████████████████| 15.2 MB 5.2 MB/s
Collecting matplotlib
  Downloading matplotlib-3.1.2-cp38-cp38-macosx_10_9_x86_64.whl (13.2 MB)
     |████████████████████████████████| 13.2 MB 4.1 MB/s
Collecting shapely
  Using cached Shapely-1.6.4.post2.tar.gz (225 kB)
Collecting pandas
  Using cached pandas-0.25.3-cp38-cp38-macosx_10_9_x86_64.whl (10.3 MB)
Collecting geopandas
  Using cached geopandas-0.6.2-py2.py3-none-any.whl (919 kB)
Collecting hvplot
  Downloading hvplot-0.5.2-py2.py3-none-any.whl (2.5 MB)
     |████████████████████████████████| 2.5 MB 4.5 MB/s
Collecting cartopy
  Using cached Cartopy-0.17.0.tar.gz (8.9 MB)
  Installing build dependencies ... done
  Getting requirements to build wheel ... error
   command: /Users/martin/anaconda3/envs/movingpandas/bin/python /Users/martin/anaconda3/envs/movingpandas/lib/python3.8/site-packages/pip/_vendor/pep517/_in_process.py get_requires_for_build_wheel /var/folders/d0/qyf0yvzd0n9ctn2wr3r4l1vw0000gn/T/tmpcf10m5i7
       cwd: /private/var/folders/d0/qyf0yvzd0n9ctn2wr3r4l1vw0000gn/T/pip-install-9up68k4a/cartopy
  Complete output (1 lines):
  Proj 4.9.0 must be installed in Conda environment "movingpandas".
ERROR: Command errored out with exit status 1: /Users/martin/anaconda3/envs/movingpandas/bin/python /Users/martin/anaconda3/envs/movingpandas/lib/python3.8/site-packages/pip/_vendor/pep517/_in_process.py get_requires_for_build_wheel /var/folders/d0/qyf0yvzd0n9ctn2wr3r4l1vw0000gn/T/tmpcf10m5i7 Check the logs for full command output.
```

Using `conda install -c anitagraser movingpandas` works, but you'll get a mixture of channels (conda-forge and defaults), which is known to [cause issues](https://geopandas.readthedocs.io/en/latest/install.html#using-the-conda-forge-channel). Moreover, you get geopandas from defaults in older version that is on conda-forge.

It is worth documenting possible issues and workarounds (strict channel priority) as the default installation instructions do not follow GeoPandas recommendations.

**Dev installation**

Current dev installation instructions expect using Anaconda. I'd say that most dev people will just make a new environment using `conda env create -f environment.yml`, but it has to be derived from the Aanconda instructions. I'd just add a sentence saying that dev env should use environment.yml and keep Anaconda instructions as one of the options only.

Minor question  - conda installs movingpandas 0.1.dev2, while the review issue says `Version submitted: 0.2`. Should I assume that 0.2 will be released after the review?
