### 2020-01-25 anitagraser/movingpandas#49 @anitagraser
Thank you for doing this review @martinfleis!

**Ad Naive installation**

Yes, pip install will also fail on Windows. My understanding is that it is not possible to pip install all GeoPandas dependencies and therefore MovingPandas suffers the same fate. I'll add Mac to the paragraph that warns Windows users.

I'll look into the channel issue. Can you please point me towards the GeoPandas recommendations that you've mentioned?

**Ad Dev installation**

I'll change the Readme accordingly.

**Ad Version in conda-forge**

I'll try to get the 0.2 release into conda-forge this weekend.
