### 2020-01-25 01417e5
Add conda instructions and Mac warning
Based on pyOpenSci review feedback https://github.com/anitagraser/movingpandas/issues/49
