### 2020-01-25 anitagraser/movingpandas#49 @anitagraser
* [x] Mac pip warning https://github.com/anitagraser/movingpandas/commit/01417e5b2ab8e3b06f8060cbcee934673ea7b4e4
* [ ] Channel issue
* [x] Dev installation instructions https://github.com/anitagraser/movingpandas/commit/01417e5b2ab8e3b06f8060cbcee934673ea7b4e4
* [x] v0.2 release on conda-forge https://anaconda.org/conda-forge/movingpandas
