### 2020-01-26 anitagraser/movingpandas#49 @martinfleis
> My understanding is that it is not possible to pip install all GeoPandas dependencies

That is true, it takes a lot of manual effort to do it.

> I'll look into the channel issue. Can you please point me towards the GeoPandas recommendations that you've mentioned?

See the note in here: https://geopandas.readthedocs.io/en/latest/install.html#using-the-conda-forge-channel

I've just realised that it is not clear that movingpandas is available on conda-forge. It might be worth it adding `conda install -c conda-forge movingpandas` next to pip install and keep it as recommended way of installation. That way most of people will succeed installation on the first try.
