### 2020-01-26 09d07a4
is_latlon compatibility with gpd 0.7
