### 2020-01-26 4d656cb
Trajectory.is_latlon() update for GeoPandas 0.7 (#50)
is_latlon compatibility with gpd 0.7
