### 2020-01-26 c84e9c0
Add conda install instructions & conda-forge link
