+++ README.md
@@ -30,11 +30,13 @@ geometries and associated attributes are stored in a GeoDataFrame *df*. Furtherm
 
 ## Installation
 
-Movingpandas for Python >= 3.7 is available from [PyPI](https://pypi.org/project/movingpandas/) and can be installed using `pip install movingpandas` 
+MovingPandas for Python >= 3.7 is available from [PyPI](https://pypi.org/project/movingpandas/) and can be installed using `pip install movingpandas`.
+
+Alternatively, MovingPandas is also available from [conda-forge](https://anaconda.org/conda-forge/movingpandas): and can be installed using `conda install -c conda-forge movingpandas`.
 
 If you're on Windows or Mac, however, many GeoPandas / MovingPandas dependencies cannot be pip installed 
 (for details see the corresponding notes in the [GeoPandas documentation](https://geopandas.readthedocs.io/en/latest/install.html#installing-with-pip)).
-Therefore, Windows users are encouraged to use the [MovingPandas conda package](https://anaconda.org/anitagraser/movingpandas). 
+Therefore, Windows users are encouraged to use the [MovingPandas conda package](https://anaconda.org/conda-forge/movingpandas).
 
 ## Development Installation 
 
