### 2020-01-26 anitagraser/movingpandas#49 @anitagraser
> It might be worth it adding conda install -c conda-forge movingpandas next to pip install and keep it as recommended way of installation.

Fixed in https://github.com/anitagraser/movingpandas/commit/c84e9c09f1bf59f35b0d995d72bfbf02b75519d5
