+++ environment.yml
@@ -1,7 +1,6 @@
 name: movingpandas
 channels:
   - conda-forge
-  - defaults
 dependencies:
   - python=3.7
   - numpy
