### 2020-01-26 537e63d
Fix slow-down due to crs check
