### 2020-01-26 ae257ca
Improve install instructions
