### 2020-01-28 7fa565d
Restore support for older envs
