### 2020-01-28 anitagraser/movingpandas#49 @anitagraser
> you'll get a mixture of channels (conda-forge and defaults), which is known to cause issues. Moreover, you get geopandas from defaults in older version that is on conda-forge.

Based on your comment, I've removed defaults from the environment. However, I've run into issues on Windows: basically, importing geopandas fails because gdal imports fail.

https://github.com/conda-forge/gdal-feedstock/issues/269#issuecomment-473661530 states that

> On Windows, because conda-forge relies on some package built with defaults blas (like scipy) one must use the defaults channel on top of conda-forge and activate conda's new strict channel feature. ... Note that on Linux and OS X we recommend the opposite order of channels.

So, I guess I'll have to add a win-environment.yml and/or add specific instructions to the README. Is there any commonly used approach for how to handle this kind of issues?
