### 2020-01-28 anitagraser/movingpandas#49 @martinfleis
> Is there any commonly used approach for how to handle this kind of issues?

I was not even aware of this Windows issue as macOS/Ubuntu user. We should probably refer to is in geopandas docs as well.

I'd say putting the content of that comment would be a good way of dealing with it.
