### 2020-01-28 9440ba1
Improve Win install instructions
