+++ environment.yml
@@ -1,6 +1,7 @@
 name: movingpandas
 channels:
   - conda-forge
+  - default
 dependencies:
   - python=3.7
   - numpy
