### 2020-01-28 anitagraser/movingpandas#51 @martinfleis
CRSError is df does not have crs
In case of creating Trajectory from GeoDataFrame without set CRS, it fails. As you moved CRS check to init, it now fails during the creation, before it failed once `is_latlon()` was called, so no big difference.

```py
df = pd.DataFrame([
  {'geometry':Point(0,0), 't':datetime(2018,1,1,12,0,0)},
  {'geometry':Point(6,0), 't':datetime(2018,1,1,12,6,0)},
  {'geometry':Point(6,6), 't':datetime(2018,1,1,12,10,0)},
  {'geometry':Point(9,9), 't':datetime(2018,1,1,12,15,0)}
]).set_index('t')
geo_df = GeoDataFrame(df)
toy_traj = mpd.Trajectory(geo_df, 1)
```

It might be expected behaviour not to allow Trajectory without set CRS. If so, I would check for missing CRS and raise more informative error. Alternatively you can assume lat-lon and raise a warning that this assumption is there.

Current `CRSError: Invalid CRS input: None` is quite easy to understand unlike previous `TypeError: ("'NoneType' object is not subscriptable", 'occurred at index 2018-01-01 12:06:00')`, but it could be pointing to a specific reason and a solution.

cc pyOpenSci/software-review#18
