### 2020-01-28 anitagraser/movingpandas#51 @martinfleis
Just realised - you might have naive coordinates without CRS, so it is probably better to assume that data are projected. For example in tracking fish movement in the tank, CRS has no meaning but movingpandas is still a tool to use.
