### 2020-01-28 anitagraser/movingpandas#52 @martinfleis
TrajectoryCollection.plot ignores some kwargs
`TrajectoryCollection.plot()` ignores some kwargs which should be passed to `GeoDataFrame.plot()`, like `figsize`.

While `Trajectory.plot(figsize=(12, 12))` works as expected, `TrajectoryCollection.plot(figsize=(12, 12))` ignores figsize. I am not sure what is the intention here, if all kwargs should go to gdf.plot() or just a selection, but difference in behaviour between Trajectory and TrajectoryCollection is unexpected.

cc pyOpenSci/software-review#18
