### 2020-01-29 anitagraser/movingpandas#53 @anitagraser
Missing badges
Necessary missing badges for pyOpenSci review:
* [x] Code coverage
* [x] pyOpenSci peer-review
* [x] repostatus.org
