### 2020-01-29 #18 @anitagraser
Thanks a lot for the thorogh review and great feedback, Martin! I'll work on the open issues.

I've been looking for the badge for pyOpenSci peer-review but haven't been able to locate one for ongoing peer review.
