### 2020-01-29 dca226e
Add PyOpenSci badge
For #53
