### 2020-01-30 2682bd1
Add codecov badge
For #53
