### 2020-01-30 anitagraser/movingpandas#53 @anitagraser
Done
