### 2020-01-30 #18 @anitagraser
If fixed the remaining README issues: badges https://github.com/anitagraser/movingpandas/issues/53 and citation information https://github.com/anitagraser/movingpandas/commit/56ef608c3302f53f408f07781fbf9c30147eb1f3

The last open issue from Martin's review should be the Community guidelines.
