### 2020-01-30 42cfe20
Create CONTRIBUTING.md
