### 2020-01-30 #18 @xmnlab
I am planning to review MovingPandas today :)
