### 2020-01-30 5fc8e8d
Add instructions for contributing
