### 2020-01-30 #18 @anitagraser
Contribution guidelines are now available at https://github.com/anitagraser/movingpandas/blob/master/CONTRIBUTING.md
