### 2020-01-30 fd9a925
Docs for TrajectoryCollectionAggregator
and add link to MyBinder
