+++ movingpandas/trajectory_aggregator.py
@@ -11,6 +11,22 @@ from .geometry_utils import azimuth, angular_difference, measure_distance_spheri
 
 class TrajectoryCollectionAggregator:
     def __init__(self, traj_collection, max_distance, min_distance, min_stop_duration, min_angle=45):
+        """
+        Create TrajectoryCollectionAggregator
+
+        Parameters
+        ----------
+        traj_collection : TrajectoryCollection
+            TrajectoryCollection to be aggregated
+        max_distance : float
+            Maximum distance between significant points
+        min_distance : float
+            Minimum distance between significant points
+        min_stop_duration : integer
+            Minimum duration required for stop detection (in seconds)
+        min_angle : float
+            Minimum angle for significant point extraction
+        """
         self.traj_collection = traj_collection
         self.max_distance = max_distance
         self.min_distance = min_distance
@@ -18,57 +34,81 @@ class TrajectoryCollectionAggregator:
         self.min_angle = min_angle
         self.is_latlon = self.traj_collection.trajectories[0].is_latlon
         print('Extracting significant points ...')
-        self.significant_points = self.extract_significant_points()
+        self.significant_points = self._extract_significant_points()
         print('  No. significant points: {}'.format(len(self.significant_points)))
         print('Clustering significant points ...')
-        self.clusters = self.cluster_significant_points()
+        self.clusters = self._cluster_significant_points()
         print('  No. clusters: {}'.format(len(self.clusters)))
         print('Computing flows ...')
-        self.flows = self.compute_flows_between_clusters()
+        self.flows = self._compute_flows_between_clusters()
         print('Flows ready!')
 
     def get_significant_points_gdf(self):
+        """
+        Return the extracted significant points
+
+        Returns
+        -------
+        GeoDataFrame
+            Significant points
+        """
         if not self.significant_points:
-            self.extract_significant_points()
+            self._extract_significant_points()
         df = DataFrame(self.significant_points, columns=['geometry'])
         return GeoDataFrame(df)
 
-    def extract_significant_points(self):
+    def get_clusters_gdf(self):
+        """
+        Return the extracted cluster centroids
+
+        Returns
+        -------
+        GeoDataFrame
+            Cluster centroids, incl. the number of clustered significant points (n)
+        """
+        if not self.clusters:
+            self._cluster_significant_points()
+        df = DataFrame([cluster.centroid for cluster in self.clusters], columns=['geometry'])
+        df['n'] = [len(cluster.points) for cluster in self.clusters]
+        return GeoDataFrame(df)
+
+    def get_flows_gdf(self):
+        """
+        Return the extracted flows
+
+        Returns
+        -------
+        GeoDataFrame
+            Flow lines, incl. the number of trajectories summarized in the flow (weight)
+        """
+        if not self.flows:
+            self._compute_flows_between_clusters()
+        return GeoDataFrame(self.flows)
+
+    def _extract_significant_points(self):
         significant_points = []
         for traj in self.traj_collection.trajectories:
-            a = PtsExtractor(traj, self.max_distance, self.min_distance, self.min_stop_duration, self.min_angle)
+            a = _PtsExtractor(traj, self.max_distance, self.min_distance, self.min_stop_duration, self.min_angle)
             significant_points = significant_points + a.find_significant_points()
         return significant_points
 
-    def cluster_significant_points(self):
+    def _cluster_significant_points(self):
         points = self.significant_points
[…]
