+++ docs/source/trajectoryaggregator.rst
@@ -1,0 +1,18 @@
+MovingPandas.TrajectoryCollectionAggregator
+===========================================
+
+.. automodule:: movingpandas
+    :members:
+    :noindex:
+
+**TrajectoryCollectionAggregator** generalizes and aggregates trajectories by extracting
+clusters of significant trajectory points and computing flows between the clusters.
+It is based on an algorithm by Andrienko & Andrienko (2011) as described in
+https://anitagraser.com/2016/11/07/movement-data-in-gis-3-visualizing-massive-trajectory-datasets/.
+
+
+.. autoclass:: TrajectoryCollectionAggregator
+   :members:
+
+   .. automethod:: __init__
+
