### 2020-01-30 886359a
Honor plot figsize, fixes #52
