### 2020-01-30 anitagraser/movingpandas#49 @anitagraser
I've added all the install details I'm aware of https://github.com/anitagraser/movingpandas/commit/9440ba13e510a0843ab248e88494afdc0a6054f2
