### 2020-01-30 anitagraser/movingpandas#52 @anitagraser
I went over the logic and `TrajectoryCollection.plot` lacked the step of creating the ax with the specified figsize. However, the rest of the logic should be same as in `Trajectory.plot`.
