### 2020-01-31 #18 @lwasser
@anitagraser when your package has fully passed both reviews and both reviewers are happy with your addressing their requested changes, we will ask you to add the badge to the readme!! please get in touch with any other questions. @martinfleis  THANK YOU for this review!!
