### 2020-01-31 8ab75d9
First paper draft
