+++ paper.bib
@@ -1,0 +1,48 @@
+@article{graser:2019,
+  title={Movingpandas: Efficient structures for movement data in python},
+  author={Graser, Anita},
+  journal={GIForum},
+  volume={1},
+  pages={54--68},
+  year={2019}
+}
+
+@software{kelsey_jordahl_2019_3545747,
+  author       = {Kelsey Jordahl and
+                  Joris Van den Bossche and
+                  Jacob Wasserman and
+                  James McBride and
+                  Martin Fleischmann and
+                  Jeffrey Gerard and
+                  Jeff Tratner and
+                  Matthew Perry and
+                  Carson Farmer and
+                  Geir Arne Hjelle and
+                  Sean Gillies and
+                  Micah Cochran and
+                  Matt Bartos and
+                  Lucas Culbertson and
+                  Nick Eubank and
+                  maxalbert and
+                  Sergio Rey and
+                  Aleksey Bilogur and
+                  Dani Arribas-Bel and
+                  Christopher Ren and
+                  Joshua Wilson and
+                  Martin Journois and
+                  Levi John Wolf and
+                  Leah Wasser and
+                  Alan D. Snow and
+                  YuichiNotoya and
+                  François Leblanc and
+                  Filipe and
+                  Chris Holdgraf and
+                  Adam Greenhall},
+  title        = {geopandas/geopandas: v0.6.2},
+  month        = nov,
+  year         = 2019,
+  publisher    = {Zenodo},
+  version      = {v0.6.2},
+  doi          = {10.5281/zenodo.3545747},
+  url          = {https://doi.org/10.5281/zenodo.3545747}
+}
