+++ paper.bib
@@ -7,6 +7,13 @@
   year={2019}
 }
 
+@article{pappalardo:2019,
+  title={scikit-mobility: An open-source Python library for human mobility analysis and simulation},
+  author={Pappalardo, Luca and Barlacchi, Gianni and Simini, Filippo and Pellungrini, Roberto},
+  journal={arXiv preprint arXiv:1907.07062},
+  year={2019}
+}
+
 @software{kelsey_jordahl_2019_3545747,
   author       = {Kelsey Jordahl and
                   Joris Van den Bossche and
