### 2020-01-31 baad2b6
More tests for TrajectoryCollection
