+++ paper.bib
@@ -1,9 +1,10 @@
 @article{graser:2019,
-  title={Movingpandas: Efficient structures for movement data in python},
+  title={MovingPandas: Efficient structures for movement data in Python},
   author={Graser, Anita},
-  journal={GIForum},
-  volume={1},
+  journal={GI\_Forum ‒ Journal of Geographic Information Science 2019},
+  volume={1-2019},
   pages={54--68},
+  doi = {10.1553/giscience2019\_01\_s54},
   year={2019}
 }
 
