+++ paper.bib
@@ -4,7 +4,7 @@
   journal={GI\_Forum ‒ Journal of Geographic Information Science 2019},
   volume={1-2019},
   pages={54--68},
-  doi = {10.1553/giscience2019\_01\_s54},
+  doi = {10.1553/giscience2019_01_s54},
   year={2019}
 }
 
