### 2020-01-31 9233b3e
Add figure
