### 2020-02-01 b7cc914
Add distance-based generalization
