### 2020-02-01 fb97166
Test generalize_min_distance
