+++ paper.bib
@@ -54,3 +54,29 @@
   doi          = {10.5281/zenodo.3545747},
   url          = {https://doi.org/10.5281/zenodo.3545747}
 }
+
+@software{philipp_rudiger_2020_3634720,
+  author       = {Philipp Rudiger and
+                  Julia Signell and
+                  James A. Bednar and
+                  Andrew and
+                  Jean-Luc Stevens and
+                  Chris B and
+                  Jordan Samuels and
+                  Todd and
+                  Thomas PEDOT and
+                  Sander van den Oord and
+                  Jon Mease and
+                  Isaac Virshup and
+                  Gabriel Corona and
+                  Danny Hermes and
+                  Curtis H. and
+                  Anita Graser},
+  title        = {holoviz/hvplot: Version 0.5.2},
+  month        = feb,
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {v0.5.2},
+  doi          = {10.5281/zenodo.3634720},
+  url          = {https://doi.org/10.5281/zenodo.3634720}
+}
