### 2020-02-04 #18 @lwasser
ok @anitagraser @arfon actually responded here: https://github.com/pyOpenSci/software-review/issues/16#issuecomment-581677530 on another review. he did say that if you have another citation already that you want to use to cite your software, then you probably don't need to worry about joss. i know you have already begun work on the joss paper... but see his comment and let us know what you think! thank you for your patience. we are still working through some of the details of our review process in these early stages
