### 2020-02-04 anitagraser/movingpandas#56 @xmnlab
[pyopensci-review] Tutorial 1: AttributeError: 'GeoDataFrame' object has no attribute 'as_matrix'
Tutorial 1, on binder, the input number 30:

```
origins['lat'] = origins.geometry.y
origins['lon'] = origins.geometry.x
matrix = origins.as_matrix(columns=['lat', 'lon'])
```

is raising the follow error:

```
<ipython-input-30-c2a4f564a8ef> in <module>
      1 origins['lat'] = origins.geometry.y
      2 origins['lon'] = origins.geometry.x
----> 3 matrix = origins.as_matrix(columns=['lat', 'lon'])

/srv/conda/envs/notebook/lib/python3.7/site-packages/pandas/core/generic.py in __getattr__(self, name)
   5271             if self._info_axis._can_hold_identifiers_and_holds_name(name):
   5272                 return self[name]
-> 5273             return object.__getattribute__(self, name)
   5274
   5275     def __setattr__(self, name: str, value) -> None:

AttributeError: 'GeoDataFrame' object has no attribute 'as_matrix'
```

ref: https://github.com/pyOpenSci/software-review/issues/18
