### 2020-02-04 anitagraser/movingpandas#57 @xmnlab
[pyopensci-review] Specify Python versions supported
As conda-forge still supports Python 3.6, probably it would be good to specify that `movingpandas` supports Python >= 3.7

ref: https://github.com/pyOpenSci/software-review/issues/18
