### 2020-02-04 #18 @xmnlab
- Minor issue: For the development installation, comment also how to install `movingpandas` for development, e.g.: `python setup.py develop` or `pip install --no-deps -e .`
  - it uses jupyter notebooks.
  - in the tutorial 0, maybe "Option 2: Creating trajectories with TrajectoryManager" should be "Option 2: Creating trajectories with TrajectoryCollection"
- [ ] **Function Documentation:** for all user-facing functions
   - There are some public functions/methods that don't have docstring. For example:
     - https://github.com/anitagraser/movingpandas/blob/fb97166309d8bd7b6f578f060a50e691c1b65cc0/movingpandas/trajectory_collection.py#L127
     - https://github.com/anitagraser/movingpandas/blob/fb97166309d8bd7b6f578f060a50e691c1b65cc0/movingpandas/trajectory_plotter.py#L65
     - https://github.com/anitagraser/movingpandas/blob/fb97166309d8bd7b6f578f060a50e691c1b65cc0/movingpandas/trajectory_plotter.py#L73
     - https://github.com/anitagraser/movingpandas/blob/fb97166309d8bd7b6f578f060a50e691c1b65cc0/movingpandas/trajectory_plotter.py#L85
     - https://github.com/anitagraser/movingpandas/blob/fb97166309d8bd7b6f578f060a50e691c1b65cc0/movingpandas/trajectory_plotter.py#L96

   You can use some tool to check the docstrings, e.g.: pydocstyle ( pydocstyle --convention=numpy ) or maybe numpydoc (https://numpydoc.readthedocs.io/en/latest/validation.html).
- [ ] **Examples** for all user-facing functions
  - Example section inside docstring function is a very good way to explain how to use it, also it will be available inside the documentation web-page. Currently, all the examples provided are inside the tutorials.
- [x] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.

Readme requirements

- [x] Badges for continuous integration and test coverage, the badge for pyOpenSci peer-review once it has started (see below), a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges, see this example, that one and that one. Such a table should be more wide than high.
- [x] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
- [x] Any additional setup required (authentication tokens, etc)
- [x] Brief demonstration usage
- [x] Direction to more detailed documentation (e.g. your documentation files or website).
- [ ] If applicable, how the package compares to other similar packages and/or how it relates to other packages
  - Not found in the README file.

  - `pip install movingpandas` didn't work
- [x] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

#### For packages co-submitting to JOSS

Estimated hours spent reviewing: 5h30

---

Review based on the version with the latest commit `fb97166`

- [ ] In the README file, maybe it would be good if `[0]` and `[1]` would be linked to the `Publications` section.
- [ ] in the "Tutorial 0", section "Continue exploring MovingPandas", maybe it would be better to have just a link to the tutorial (1).
- [ ] in the "Tutorial 0", probably "Option 2: Creating trajectories with TrajectoryManager" should be "Option 2: Creating trajectories with TrajectoryCollection"
- [ ] the "Tutorial 3", doesn't have a section "Continue exploring MovingPandas".
- [ ] the installation using pip failed.
- [ ] explain how to install `movingpandas` in development mode.
- [ ] add missing docstrings for user-facing functions/methods.
- [ ] add `Examples` section in the user-facing functions/methods docstrings.
- [ ] add into README file how `movingpandas` compares to other similar packages
- [ ] the title of the tutorials 3 and 4 doesn't match to the name provided in the tutorial list in the "Tutorial 0"
- [ ] check why `pip install movingpandas` didn't work

Suggestions (optional):

- on CI, check if the code fit in the PEP 8 style (for example, using flake8).
- it would be good to have all tutorials statically inside the documentation page (eg. https://pandas.pydata.org/pandas-docs/stable/getting_started/basics.html, https://docs.ibis-project.org/tutorial.html)
- also each static tutorial page could have a link to mybinder (eg. https://sunpy.org/posts/2018/2018-07-21-coronal-loop-coordinates)
- as it is still common to have Python 3.6 environment and it is not supported by `movingpandas`, it would be good to comment in the documenation the Python versions supported.
- CI maybe could cover the MacOS and MS Windows. Also it could cover all Python version supported.
- in the tutorials, add to the charts a title, a description for each axis and its respectively unit and legends.
- in the "Tutorial 4", add the description about this notebook.
