### 2020-02-04 #18 @xmnlab
@anitagraser it is a interesting project!

I have added the result of my review above. Let me know if you need any additional explanation or detail!
