### 2020-02-06 #18 @anitagraser
Thank you very much for the extensive feedback @xmnlab! It's very clear. I have just one question so far concerning "explain how to install movingpandas in development mode.": the [README](https://github.com/anitagraser/movingpandas#development-installation) already has a section titled Development Installation with the necessary conda instructions. Are you missing something specific?
