### 2020-02-06 49e4c59
Replaced deprecated as_matrix call, fixes #56
