### 2020-02-06 16831fd
Specify Python version, fixes #57
