### 2020-02-06 eb33d4d
Explain pip issues, fixes #58
