### 2020-02-06 #18 @anitagraser
Status of improvements:

* [x] In the README file, maybe it would be good if [0] and [1] would be linked to the Publications section. https://github.com/anitagraser/movingpandas/commit/f1da7954563f79507ea8e62222bdf677ca135d13
* [x]  in the "Tutorial 0", section "Continue exploring MovingPandas", maybe it would be better to have just a link to the tutorial (1).  https://github.com/anitagraser/movingpandas/commit/1b1ee2ec8d9461a698a8fe24380da8e2ef2f77f3
* [ ] in the "Tutorial 0", probably "Option 2: Creating trajectories with TrajectoryManager" should be "Option 2: Creating trajectories with TrajectoryCollection"
* [x]  the "Tutorial 3", doesn't have a section "Continue exploring MovingPandas". https://github.com/anitagraser/movingpandas/commit/1b1ee2ec8d9461a698a8fe24380da8e2ef2f77f3
* [x]  the installation using pip failed. https://github.com/anitagraser/movingpandas/commit/eb33d4d7eef4cd79640473380d76bb99ad9e8a46
* [ ]  explain how to install movingpandas in development mode.
* [ ]  add missing docstrings for user-facing functions/methods.
* [ ]  add Examples section in the user-facing functions/methods docstrings.
* [ ]  add into README file how movingpandas compares to other similar packages
* [ ]  the title of the tutorials 3 and 4 doesn't match to the name provided in the tutorial list in the "Tutorial 0" https://github.com/anitagraser/movingpandas/commit/1b1ee2ec8d9461a698a8fe24380da8e2ef2f77f3
* [x]  check why pip install movingpandas didn't work https://github.com/anitagraser/movingpandas/commit/eb33d4d7eef4cd79640473380d76bb99ad9e8a46
