### 2020-02-08 a0607d1
Add missing docstring and mark plotter classes private
