### 2020-02-08 7f2047f
More usage examples
