### 2020-02-10 #18 @jlpalomino
@anitagraser thanks for checking in about this and for your action! We are now working on a PR to the dev guide (https://github.com/pyOpenSci/dev_guide/pull/42) to address the timing of when to add the badge as well as to include the badge information for when the package is accepted.
