### 2020-02-10 #18 @xmnlab
@anitagraser about the "Development Installation" it explains how to prepare the environment but I didn't find the command line for install `movingpandas` in development mode, it could be, for example, `python setup.py develop` or `pip install --no-deps -e .` or if you want to just point to some nice documentation/tutorial that explains more about development mode  you can add just a link to that page (eg.: https://python-packaging-tutorial.readthedocs.io/en/latest/setup_py.html#develop-mode) or also you can do both
