### 2020-02-16 897e82c
Fix plot speed for trajectory collections
