### 2020-02-17 #18 @xmnlab
@anitagraser it looks very nice!

related to your point about the methods self-explanatory, I would add examples for each function/method  including for these cases because:

1) if you want you can also integrate it with doctest (https://docs.python.org/3/library/doctest.html) in the future and it will help to test your docstring examples and
2) for some functions that return more complex data (like dataframe or series) it helps to understand better the data returned (also maybe it would be good to add more details about the data structure (eg: columns names and data types) for functions that return dataframes in the `Returns` section.

@lwasser do you have any thoughts about this point?
