### 2020-02-17 #18 @xmnlab
@martinfleis actually I am very curious about this too :) as the guideline says that `all facing-user function should have examples` (https://www.pyopensci.org/dev_guide/packaging/packaging_guide.html#Documentation) so using just tutorials for the examples is hard to check if all functions have an example inside. also from the user's perspective, If I want to see how to use a method, generally I will go to the API documentation not the tutorials.

examples inside the docstring will be published automatically inside the documentation (if the project generate the documentation from the code).

maybe I didn't express me well about the doctest ... I didn't want to suggest to use that .. I just want to comment that also it could help in the future .. but it is not required. the main point about the examples was the item 2.

But as required by https://www.pyopensci.org/dev_guide/packaging/packaging_guide.html#Documentation my first guess is that all public functions should have an example inside the docstrings or on the API documentation.

@jlpalomino could you guide us about this topic?
