### 2020-02-17 #18 @anitagraser
Thank you all for working on clarifying this! I mostly looked at previously accepted projects to guide my work. For example, Pandera has usage examples but not for all user-facing functions:  https://pandera.readthedocs.io/en/latest/API.html
