### 2020-02-17 #18 @jlpalomino
@anitagraser @martinfleis @xmnlab This is a great discussion, and I want to encourage you to continue it.  I am interjecting here to say that I will discuss these comments further with @lwasser and get back to you with feedback from the editor perspective.
