### 2020-02-18 #18 @martinfleis
When I started my review I tried to understand what pyOpenSci expects. Based on the review guide, I felt that it tries to be rather inclusive than strict. After reading various Good/Better/Best explanations, my understanding was following:

- The package does not have to be _perfect_. Various Good options are far from perfect (e.g. `Manually updated documentation as text files that ship with your package.`), but they are good enough to send the package to the world.
- As a reviewer, I should try to find issues. Those which normal user might face and those developer might face. If they are significant, they should be fixed. If they are not significant or will become significant in a year (deprecation etc.), it is fine to point them out and leave developer to fix them later.
- I should make sure that every aspect of the package is at least at a Good level. Then I can make recommendations how to get to Better or Best level, but these, again, should not be in a way of acceptance.
- If everything is at least at a Good level and I gave my recommendations how to make it even better, I should approve the package for pyOpenSci publication.

These were (roughly) the rules I applied to my review. I would be happy to hear from @lwasser and @jlpalomino if they fit the aim of pyOpenSci. How strict do you want to be in the review?
