### 2020-02-20 #18 @jlpalomino
Hi @anitagraser @martinfleis @xmnlab thanks for your patience in this process.

@lwasser and I have decided that this is a good topic of conversation during the next pyOpenSci meeting, which will occur next Fri 2/28/20 at 11am MDT ([details on discourse](https://pyopensci.discourse.group/t/our-next-community-call-is-fri-28-feb-2020-at-11am-mountain-time/158/2)). We invite you all to attend if you are able.
