### 2020-02-20 #18 @jlpalomino
We will also start a discourse discussion on this!
