### 2020-03-01 #18 @anitagraser
Thank you for the update, @lwasser ! If I understand the intent of the discussion on discourse, having examples for all user-facing functions is not going to be a must.
Besides trying to find the best wording for the guidelines, how should we proceed on this review?
If I haven't missed anything, my work on the requested improvements is finished.
