### 2020-03-01 #18 @anitagraser
@martinfleis Right! For JOSS, I'll still have to get a DOI.

I might bump the MovingPandas version to 0.3. (Movingpandas 0.2 works for geopandas 0.6.3. MovingPandas 0.3 works with GeoPandas 0.7)

The paper is in a dedicated brach for now: https://github.com/anitagraser/movingpandas/blob/joss-paper/paper.md

Is there anything else?
