### 2020-03-01 9d005c3
Fix to_crs needs to update is_latlon
