### 2020-03-01 388a2cf
Pin bokeh<2.0
