### 2020-03-02 #18 @martinfleis
I've checked the paper and there does not seem to be any issue re JOSS requirements. Just few very minor notes.

- matplotlib reference is missing
- after MyBinder link is missing dot
- JOSS apparently makes a new line in the middle of `code`, which in some cases breaks movingpandas, geopandas,  and python words. See linked pdf - [paper.pdf](https://github.com/pyOpenSci/software-review/files/4277970/qwex64zqrahji9pkdssm.pdf). You can check how it renders on https://whedon.theoj.org.

I'll formally update my review post to include JOSS boxes, but I am happy with the package and all responses and recommend it for a publication on both pyOpenSci and JOSS.
