### 2020-03-02 d2e835c
Add missing ref and missing dot
As pointed out in https://github.com/pyOpenSci/software-review/issues/18#issuecomment-593624938
