### 2020-03-07 #18 @anitagraser
@martinfleis Thank you for reviewing the paper! I have fixed the first two issues. I don't see how to affect the line break (third issue you mentioned) and found that it's also visible in recently published papers, such as https://joss.theoj.org/papers/10.21105/joss.02075
