### 2020-03-09 #18 @jlpalomino
Hi @anitagraser, it is correct that based on the recent pyOpenSci guideline discussion, examples for all user-facing functions would not be considered a requirement under the OK/Better/Best framework. For the overall review, if the reviewers felt that the `OK` requirements have been met and recommend accepting the package, then a package could be accepted to pyOpenSci.

For everyone - we welcome all additional comments on the [OK/Better/Best framework and/or the overall acceptance process]( https://pyopensci.discourse.group/t/good-better-best-required-versus-recommended-changes-by-reviewers-your-feedback-is-requested/160/2).

Thanks @martinfleis for expanding your review to include the JOSS tick boxes.

@xmnlab Can you please expand your review to address the tick boxes for the JOSS submission? The paper can be found at:  https://github.com/anitagraser/movingpandas/blob/joss-paper/paper.md
Thanks!
