### 2020-03-12 #18 @xmnlab
@anitagraser thanks for having working hard on all the recommendations.

MovingPandas documenation looks very good. I just found some minor issues. Some methods' docstring don't have the parameter section:

MovingPandas.TrajectoryCollection:
- `TrajectoryCollection.add_speed` https://movingpandas.readthedocs.io/en/latest/trajectorycollection.html#movingpandas.TrajectoryCollection.add_speed

MovingPandas.Trajectory
- `Trajectory.add_direction` https://movingpandas.readthedocs.io/en/latest/trajectory.html#movingpandas.Trajectory.add_direction
- `Trajectory.add_speed` https://movingpandas.readthedocs.io/en/latest/trajectory.html#movingpandas.Trajectory.add_speed
- `Trajectory.to_crs` https://movingpandas.readthedocs.io/en/latest/trajectory.html#movingpandas.Trajectory.to_crs

I am starting to review JOSS section now. thanks again!
