### 2020-03-12 #18 @xmnlab
Some suggestions related to docstrings (not required):
- Inside `Parameters` section, you can use `optional` and `default` information to add more details about the parameter, for example:

```
x : int, optional
    Some description of parameter `x` (the default is -1, which implies summation over all axes).
```
ref: https://numpydoc.readthedocs.io/en/latest/format.html
- When a parameter can only assume one of a fixed set of values, those values can be listed in braces, with the default appearing first, for example:  `order : {'C', 'F', 'A'}`
