### 2020-03-12 b112323
Add panel
Needed for panel demo notebook
