### 2020-03-12 #18 @xmnlab
JOSS Review section:

It looks very good! I just noticed a minor issue, also mentioned by  @martinfleis , related to some word break like GeoPandas, MovingPandas.

thanks for working on that @anitagraser , let me know if I can help in anyway.
