### 2020-03-14 #18 @anitagraser
> MovingPandas documenation looks very good. I just found some minor issues. Some methods' docstring don't have the `parameters` section:

Thank you, @xmnlab! I've added the missing parameter sections: https://github.com/anitagraser/movingpandas/commit/b98b404238e6880f22c3d8743d45a1cbad1d3a34
