### 2020-03-14 9081cea
Bump to v0.3.rc1
0.3 includes GeoPandas 0.7 support (new CRS handling)
