### 2020-03-14 #18 @anitagraser
And a Zenodo DOI is set up [![DOI](https://zenodo.org/badge/161995245.svg)](https://zenodo.org/badge/latestdoi/161995245)
