### 2020-03-15 897f8b6
Pin bokeh<2.0
Fixes #60
