### 2020-03-16 #18 @anitagraser
v0.3rc1 has landed on conda-forge https://anaconda.org/conda-forge/movingpandas
