### 2020-03-16 #18 @xmnlab
awesome! @anitagraser thanks for letting us know about the new release :)

Final approval (post-review)
- [x] The author has responded to my review and made changes to my satisfaction. I recommend approving this package.

Thanks @anitagraser for working on the recommendations and congrats for the hard work you are doing on MovingPandas!
