### 2020-03-16 #18 @jlpalomino
Thank you @martinfleis and @xmnlab for your final reviews of MovingPandas.  We also really appreciate all of your feedback on the review process.

Great work @anitagraser!  Both reviewers have provided their final approvals for acceptance to pyOpenSci.  I will follow up with next steps shortly.  Thanks.
