+++ movingpandas/overlay.py
@@ -1,210 +1,0 @@
-# -*- coding: utf-8 -*-
-
-import pandas as pd
-from shapely.geometry import Point, LineString, shape
-from shapely.affinity import translate
-from datetime import datetime, timedelta
-
-
-class SpatioTemporalRange:
-    def __init__(self, pt_0, pt_n, t_0, t_n):
-        self.pt_0 = pt_0
-        self.pt_n = pt_n
-        self.t_0 = t_0
-        self.t_n= t_n
-
-
-class TemporalRange:
-    def __init__(self, t_0, t_n):
-        self.t_0 = t_0
-        self.t_n= t_n
-
-
-def _get_spatiotemporal_ref(row):
-    """Returns the SpatioTemporalRange for the input row's spatial_intersection LineString
-    by interpolating timestamps.
-    """
-    if type(row['spatial_intersection']) == LineString:
-        pt0 = Point(row['spatial_intersection'].coords[0])
-        ptn = Point(row['spatial_intersection'].coords[-1])
-        t = row['prev_t']
-        t_delta = row['t'] - t
-        length = row['line'].length
-        t0 = t + (t_delta * row['line'].project(pt0)/length)
-        tn = t + (t_delta * row['line'].project(ptn)/length)
-        # to avoid numerical issues with microseconds beyond six digits, we reconstruct the timestamps
-        t0 = datetime(t0.year, t0.month, t0.day, t0.hour, t0.minute, t0.second, t0.microsecond)
-        tn = datetime(tn.year, tn.month, tn.day, tn.hour, tn.minute, tn.second, tn.microsecond)
-        # to avoid intersection issues with zero length lines
-        if ptn == translate(pt0, 0.00000001, 0.00000001):
-            t0 = row['prev_t']
-            tn = row['t']
-        # to avoid numerical issues with timestamps
-        if is_equal(tn, row['t']):
-            tn = row['t']
-        if is_equal(t0, row['prev_t']):
-            t0 = row['prev_t']
-        return SpatioTemporalRange(pt0, ptn, t0, tn)
-    else:
-        return None
-
-
-def _dissolve_ranges(ranges):
-    """SpatioTemporalRanges that touch (i.e. the end of one equals the start of another) are dissolved (aka. merged)."""
-    if len(ranges) == 0:
-        raise ValueError("Nothing to dissolve (received empty ranges)!")
-    new = []
-    start = None
-    end = None
-    pt0 = None
-    ptn = None
-    for r in ranges:
-        if r is None:
-            continue  # raise ValueError('Received range that is None!')
-        if start is None:
-            start = r.t_0
-            end = r.t_n
-            pt0 = r.pt_0
-            ptn = r.pt_n
-        elif end == r.t_0:
-            end = r.t_n
-            ptn = r.pt_n
-        elif r.t_0 > end and is_equal(r.t_0, end):
-            end = r.t_n
-            ptn = r.pt_n
-        else:
-            new.append(SpatioTemporalRange(pt0, ptn, start, end))
-            start = r.t_0
-            end = r.t_n
-            pt0 = r.pt_0
-            ptn = r.pt_n
-    new.append(SpatioTemporalRange(pt0, ptn, start, end))
-    return new
-
-
-def is_equal(t1, t2):
-    """Similar timestamps are considered equal to avoid numerical issues."""
-    if type(t2) == pd.Timestamp:
-        td = abs(t1 - t2.tz_localize(t1.tzinfo).to_pydatetime())
-    else:
-        td = abs(t1 - t2)
-    return td < timedelta(milliseconds=10)
-
-
-def intersects(traj, polygon):
-    try:
-        line = traj.to_linestring()
-    except:
-        return False
[…]
