+++ movingpandas/__init__.py
@@ -1,10 +1,0 @@
-"""
-`movingpandas`: Implementation of Trajectory classes and functions built on top of GeoPandas
-"""
-
-from .trajectory import Trajectory
-from .trajectory_collection import TrajectoryCollection
-from .trajectory_aggregator import TrajectoryCollectionAggregator
-
-name = 'movingpandas'
-__version__ = '0.2.rc2'
