+++ _config.yml
@@ -1,0 +1,3 @@
+heme: jekyll-theme-cayman
+title: MovingPandas 
+description: MovingPandas implements a Trajectory class and corresponding methods based on GeoPandas.
