### 2020-03-19 anitagraser/movingpandas#49 @jlpalomino
Hi @martinfleis. Trying to clean loose ends in this review.  It looks like @anitagraser has indicated that this issue has been addressed (and you indicated a thumbs up :). Can this issue be closed?
