### 2020-03-19 anitagraser/movingpandas#51 @jlpalomino
Hi @martinfleis it looks like @anitagraser has flagged this issue as a future enhancement.  Are you comfortable with the labeling of this issue?  (It would remain as the last open issue from this review). Thanks.
