### 2020-03-19 #18 @jlpalomino
Hi @anitagraser you may have seen that I have followed up with @martinfleis to tie up loose ends related to two open issues from his review.

One item is for you is about versions. The last JOSS check for the editor is: `Version (JOSS only, may be post-review): Does the release version given match the GitHub release (v1.0.0)?`

It seems that there is a slight difference between the [GitHub version 0.3.rc2](https://github.com/anitagraser/movingpandas/blob/master/setup.py) versus [Zenodo version v0.3.rc1](https://zenodo.org/record/3710951) and [conda forge version v0.3.rc1](https://anaconda.org/conda-forge/movingpandas).  Please let me know if I am misinterpreting this information.  Thanks!
