### 2020-03-19 #18 @anitagraser
Thank you @jlpalomino!

Concerning versions: in my experience, it is common practice to increase the version in Github immediately after a release. Think of it as a necessary step for starting work on the next release.
