### 2020-03-19 anitagraser/movingpandas#49 @martinfleis
@jlpalomino I guess so, I am fine with the current solution. I'll close it and @anitagraser can reopen if needed.
