### 2020-03-19 anitagraser/movingpandas#51 @martinfleis
@jlpalomino Yes. It is not blocking the review.
