### 2020-03-19 #18 @martinfleis
@jlpalomino I am happy with all changes. I thought I indicated it clearly above, but apparently not enough. There is nothing to be resolved from my side.

@anitagraser versions should be ideally in sync between GitHub, PyPI, conda-forge and zenodo.
