### 2020-03-19 anitagraser/movingpandas#51 @jlpalomino
Thank you for the prompt response @martinfleis!
