### 2020-03-19 anitagraser/movingpandas#49 @jlpalomino
Again thanks for confirming @martinfleis!
