+++ LICENSE.txt
@@ -1,6 +1,6 @@
 BSD 3-Clause License
 
-Copyright (c) 2018-2019, MovingPandas developers
+Copyright (c) 2018-2020, MovingPandas developers
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
