### 2020-03-19 #18 @anitagraser
I've reverted the version number to rc1 https://github.com/anitagraser/movingpandas/commit/bba76fe9fbdc4a40a8f6aa8054046171c3b28190
