### 2020-03-19 #18 @jlpalomino
Thanks @martinfleis for your prompt responses about those issues (for final documentation purposes) and for your response related to the versions.

@anitagraser thanks for your action on the version. I checked in with @lwasser about the version difference as well. We have typically followed the version sync suggested by @martinfleis, that the GitHub repository remains the same version and only changes when there is a new release.  However, we understand that there could be commits happening before the new release.  She has suggested that we open a discussion on discourse to see what others think is best practice for versions.

That said, let's start moving forward with closing this review, while we wait for the community to weigh in. I will post a new comment with the next steps.
