### 2020-03-19 #18 @jlpalomino
MovingPandas has been approved for pyOpenSci! Thanks @anitagraser for this submission and @martinfleis and @xmnlab for your detailed reviews!

@anitagraser Here are the next steps:
- [ ] Add the badge for pyOpenSci peer review, with the second link being the URL to this issue:
[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/18)
- [ ] Submit a PR to `pyopensci.github.io` to add MovingPandas to the pyOpenSci [package list](https://github.com/pyOpenSci/pyopensci.github.io/blob/master/_data/packages.yml) and to add yourself as a [contributor](https://github.com/pyOpenSci/pyopensci.github.io/blob/master/_data/contributors.yml) using:
```
contributor_type:
    - package-maintainer
```
and
```
packages-submitted: ["movingpandas"]
```

If you are interested (not required), you can write a blog post for the pyOpenSci website about MovingPandas (see blog post about [pandera](https://www.pyopensci.org/blog/pandera-python-pandas-dataframe-validation)) to promote your package.

The last action item is for me to get the process started with JOSS, who will provide more information on their process. I will do that in a new comment.

Please feel free to let me know if you have any questions, and congrats again.
