### 2020-03-19 #18 @jlpalomino
Hi @arfon pyOpenSci has approved MovingPandas. @anitagraser is interested in JOSS publication, and the [draft paper](https://github.com/anitagraser/movingpandas/blob/joss-paper/paper.md) has been reviewed by the pyOpenSci reviewers.

@anitagraser has another [publication](https://www.austriaca.at/rootcollection?arp=0x003aba2b) that she feels demonstrates minimal overlap:
Graser, A. (2019). MovingPandas: Efficient Structures for Movement Data in Python. GI_Forum ‒ Journal of Geographic Information Science 2019, 1-2019, 54-68. doi:10.1553/giscience2019_01_s54.

Please feel free to contact @anitagraser (or myself if needed) for any additional information needed for the JOSS review process. Thanks!
