### 2020-03-20 90b3e81
Add pyOpenSci badge
Granted in https://github.com/pyOpenSci/software-review/issues/18#issuecomment-601421842
