### 2020-03-20 #18 @arfon
@anitagraser - feel free to open an issue on the [JOSS repository](https://github.com/openjournals/joss/issues) about this paper. On initial inspection I would say that JOSS would not accept a paper about MovingPandas as the [earlier publication](https://www.austriaca.at/rootcollection?arp=0x003aba2b) looks to be describing essentially the same software.
