### 2020-03-20 #18 @anitagraser
Thank you for your feedback, @arfon!
As mentioned in https://github.com/pyOpenSci/software-review/issues/18#issuecomment-580630767, I wasn't sure if prior publications would be an obstacle. Graser (2019) describes the concepts underlying MovingPandas and presents an early unreleased version of the library. (The first release was published later in Sept 2019 https://pypi.org/project/movingpandas/#history.) MovingPandas has evolved considerably since then.
If you think that this is clearly against JOSS requirements, I won't purse it further.
