### 2020-03-20 #18 @arfon
@anitagraser - could you summarize the major changes between MovingPandas when Graser (2019) was submitted compared to how it is today?

We do allow for multiple papers for the same piece of software but would expect _at least_ a major release of the software to warrant an additional paper.
