### 2020-03-20 #18 @anitagraser
@arfon Thank you for the clarification!

The key improvements since Graser (June 2019) are:

* New data structure
   * [TrajectoryCollection](https://github.com/anitagraser/movingpandas/blob/master/movingpandas/trajectory_collection.py) class: handles sets of Trajectories and enables convenient exploration and easier access to individual trajectories
* New functionality
   * [Plotting of Trajectories and TrajectoryCollections](https://github.com/anitagraser/movingpandas/blob/master/movingpandas/trajectory_plotter.py)
      * Static plots using Matplotlib
      * Interactive plots using Holoviews
   * [Aggregations for TrajectoryCollections](https://github.com/anitagraser/movingpandas/blob/master/movingpandas/trajectory_aggregator.py)
* New project infrastructure
   * Releases on pypi and conda-forge
   * Docs on readthedocs.io
   * [Extensive tutorial notebooks](https://mybinder.org/v2/gh/anitagraser/movingpandas/binder-tag?filepath=tutorials/0_getting_started.ipynb)
   * Code coverage monitoring
   * Automatic testing with Travis CI
