### 2020-03-22 26cdf94
Clean up home page
