+++ _config.yml
@@ -1,3 +1,3 @@
 theme: jekyll-theme-cayman
 title: MovingPandas 
-description: MovingPandas implements a Trajectory class and corresponding methods based on GeoPandas.
+description: Movement data exploration and analysis
