### 2020-03-31 #18 @anitagraser
@arfon Do you suggest moving this over to the JOSS repository?
