### 2020-03-31 f7ee134
Add support for Point subclasses, fixes #63
