### 2020-03-31 13304fc
Add support for custom geometry column names, fixes #62
