+++ movingpandas/tests/test_trajectory.py
@@ -296,8 +296,21 @@ class TestTrajectory:
         ]).set_index('t')
         geo_df = GeoDataFrame(df, crs=CRS_METRIC)
         traj = Trajectory(geo_df, 1)
+        traj.add_speed()
+        traj.add_direction()
         traj.hvplot()
 
+    def test_support_for_other_geometry_column_names(self):
+        df = pd.DataFrame([
+            {'xxx': TestPoint(0, 0), 't': datetime(2018, 1, 1, 12, 0, 0)},
+            {'xxx': TestPoint(6, 0), 't': datetime(2018, 1, 1, 12, 6, 0)},
+            {'xxx': TestPoint(6, 6), 't': datetime(2018, 1, 1, 12, 10, 0)}
+        ]).set_index('t')
+        geo_df = GeoDataFrame(df, geometry='xxx', crs=CRS_METRIC)
+        traj = Trajectory(geo_df, 1)
+        traj.add_speed()
+        traj.add_direction()
+        traj.hvplot()
 
     """ 
     This test should work but fails in my PyCharm probably due to https://github.com/pyproj4/pyproj/issues/134
