### 2020-03-31 7c2b7a5
Add support for other timestamp column names, fixes #61
