### 2020-03-31 5e73a53
Remove bokeh pin
Travis started failing because geoviews needs bokeh>=2.0
