### 2020-04-01 #18 @arfon
Hi @anitagraser and @lwasser, apologies for the slow reply here. @anitagraser - thanks for documenting the changes that you've made to this package since the last paper was published.

After giving this some thought, I don't think we would accept this as a JOSS submission. Partly because of the reason that @lwasser mentions (infrastructure improvements are obviously useful/important but not actually new functionality in the software) but also that the changes in the `TrajectoryCollection` and supporting classes looks to be relatively modest and should they be part of a standalone package would likely fall into our 'minor utility' category.

> This is interesting because we need to define what is a major improvement.

I completely agree. Unfortunately JOSS doesn't have docs to help with this. I'd be open to having a broader discussion to try and define this.

> We want to ensure that software publication via JOSS is robust and there is not duplication of ideas published.

Thank you, it's much appreciated and sorry again for the delayed response here.
