### 2020-04-02 66e710f
support custom geometry column name support dataframes with additional columns
