### 2020-04-02 74645fa
Additional fixes for #62 (#65)
* support custom geometry column name
support dataframes with additional columns

* revert unnecessary test
