+++ movingpandas/trajectory_aggregator.py
@@ -28,6 +28,10 @@ class TrajectoryCollectionAggregator:
             Minimum angle for significant point extraction
         """
         self.traj_collection = traj_collection
+        if self.traj_collection.trajectories:
+            self._crs = traj_collection.trajectories[0].crs
+        else:
+            self._crs = None
         self.max_distance = max_distance
         self.min_distance = min_distance
         self.min_stop_duration = min_stop_duration
@@ -55,7 +59,7 @@ class TrajectoryCollectionAggregator:
         if not self.significant_points:
             self._extract_significant_points()
         df = DataFrame(self.significant_points, columns=['geometry'])
-        return GeoDataFrame(df)
+        return GeoDataFrame(df, crs=self._crs)
 
     def get_clusters_gdf(self):
         """
@@ -70,7 +74,7 @@ class TrajectoryCollectionAggregator:
             self._cluster_significant_points()
         df = DataFrame([cluster.centroid for cluster in self.clusters], columns=['geometry'])
         df['n'] = [len(cluster.points) for cluster in self.clusters]
-        return GeoDataFrame(df)
+        return GeoDataFrame(df, crs=self._crs)
 
     def get_flows_gdf(self):
         """
@@ -83,7 +87,7 @@ class TrajectoryCollectionAggregator:
         """
         if not self.flows:
             self._compute_flows_between_clusters()
-        return GeoDataFrame(self.flows)
+        return GeoDataFrame(self.flows, crs=self._crs)
 
     def _extract_significant_points(self):
         significant_points = []
@@ -124,7 +128,7 @@ class _PtsExtractor:
     def find_significant_points(self):
         while self.j < self.n - 1:
             if self.is_representative_max_distance():
-                self.significant_points.append(self.traj.df.iloc[self.j].geometry)
+                self.significant_points.append(self.traj.df.iloc[self.j][self.traj.get_geom_column_name()])
                 self.i = self.j
                 self.j = self.i + 1
                 continue
@@ -133,7 +137,7 @@ class _PtsExtractor:
                     d_time = self.traj.df.iloc[self.k - 1].name - self.traj.df.iloc[self.j].name
                     if d_time >= self.min_stop_duration:
                         # print("significant stop ({1}) at {0}".format(self.j,d_time))
-                        self.significant_points.append(self.traj.df.iloc[self.j].geometry)
+                        self.significant_points.append(self.traj.df.iloc[self.j][self.traj.get_geom_column_name()])
                         self.i = self.j
                         self.j = self.k
                         continue
@@ -144,7 +148,7 @@ class _PtsExtractor:
                 a_turn = self.compute_angle_between_vectors()
                 if a_turn >= self.min_angle and a_turn <= (360 - self.min_angle):
                     # print("significant angle ({0}) at {1}".format(a_turn,self.j))
-                    self.significant_points.append(self.traj.df.iloc[self.j].geometry)
+                    self.significant_points.append(self.traj.df.iloc[self.j][self.traj.get_geom_column_name()])
                     self.i = self.j
                     self.j = self.k
                 else:
@@ -154,17 +158,17 @@ class _PtsExtractor:
         return self.significant_points
 
     def compute_angle_between_vectors(self):
-        p_i = self.traj.df.iloc[self.i].geometry
-        p_j = self.traj.df.iloc[self.j].geometry
-        p_k = self.traj.df.iloc[self.k].geometry
+        p_i = self.traj.df.iloc[self.i][self.traj.get_geom_column_name()]
[…]
