### 2020-04-05 96d4de5
Refactor trajectory generalization
