### 2020-04-05 128a097
More generalizer refactoring
