+++ movingpandas/__init__.py
@@ -3,7 +3,7 @@
 """
 
 from .trajectory import Trajectory
-from .trajectory_generalizer import TrajectoryGeneralizer
+from .trajectory_generalizer import TrajectoryGeneralizer, DouglasPeuckerGeneralizer, MinDistanceGeneralizer, MinTimeDeltaGeneralizer
 from .trajectory_collection import TrajectoryCollection
 from .trajectory_aggregator import TrajectoryCollectionAggregator
 
