+++ docs/source/trajectorygeneralizer.rst
@@ -1,0 +1,21 @@
+MovingPandas.TrajectoryGeneralizer
+===========================================
+
+.. automodule:: movingpandas
+    :members:
+    :noindex:
+
+.. autoclass:: TrajectoryGeneralizer
+   :members:
+
+   .. automethod:: __init__
+
+.. autoclass:: DouglasPeuckerGeneralizer
+   :members:
+
+.. autoclass:: MinDistanceGeneralizer
+   :members:
+
+.. autoclass:: MinTimeDeltaGeneralizer
+   :members:
+
