### 2020-04-07 #18 @lwasser
@arfon  thank you . we understand. and we'd be happy to participate in a discussion surrounding what defines a major improvement. It will make things simpler for future reviews as we can provide that information to the maintainer at the beginning if they are interested in JOSS as well!

@anitagraser i think we can close this issue. Are you interested / do you have the time to write a  blog post - similar to what Niels did for pandera:

https://www.pyopensci.org/blog/pandera-python-pandas-dataframe-validation

no pressure as i know there is a lot going on now AND you are welcome to take your time doing it. We will then put the word about about your package and link to the post which describes it in more detail! let me know what you think!
