### 2020-04-08 #18 @jlpalomino
@martinfleis We noticed that you are not listed on our contributors page. When you get a chance, can you please submit a PR to `pyopensci.github.io` to add yourself as a [contributor](https://github.com/pyOpenSci/pyopensci.github.io/blob/master/_data/contributors.yml) using:
```
contributor_type:
    - reviewer
```
