### 2020-04-08 #18 @jlpalomino
Thanks everyone!  I am officially closing this issue.

@anitagraser feel free to reach out about a blog post on MovingPandas if you are interested!
